+++ libbeat/processors/ratelimit/rate_limit_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package rate_limit
+package ratelimit
 
 import (
 	"testing"
